Refactor channel interactions to enable usage in E2E tests
